Extract method generators to separate classes
What?
=====

This extracts all of the method generation behavior to discrete classes
for each of the has_one, has_many, and has_many_ordered behaviors.
