+++ lib/page_ez.rb
@@ -7,6 +7,12 @@ require_relative "page_ez/null_logger"
 require_relative "page_ez/delegates_to"
 require_relative "page_ez/parameters"
 require_relative "page_ez/selector_evaluator"
+require_relative "page_ez/method_generators/has_one_static_selector"
+require_relative "page_ez/method_generators/has_one_dynamic_selector"
+require_relative "page_ez/method_generators/has_one_composed_class"
+require_relative "page_ez/method_generators/has_one_predicate_methods"
+require_relative "page_ez/method_generators/has_many_ordered_selector"
+require_relative "page_ez/method_generators/has_many_static_selector"
 require_relative "page_ez/visitors/matcher_collision_visitor"
 require_relative "page_ez/visitors/depth_visitor"
 require_relative "page_ez/visitors/debug_visitor"
